tests/wpt/web-platform-tests/app-history/navigate/goTo-in-detached-window.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(async () => {
    let iframe_constructor = i.contentWindow.DOMException;
    let i_win = i.contentWindow;
    let key = i_win.appHistory.current.key;
    i.remove();
    await promise_rejects_dom(t, 'InvalidStateError', iframe_constructor, i_win.appHistory.goTo(key));
    t.done();
  });
}, "goTo() in a detached window");
</script>
